Update for network_tester application
The tester application can now also measure the message latency of
payload messages. The user can define the number of messages to be sent
per handshake connection. The message size is also user configurable.

The current loop logic is very complex and dirty. This definitely needs
a refactoring in the future.